move file into docker folder
