+++ mydockers/sshmitm/src/sshmitm.py
@@ -1,22 +1,61 @@
 
 from binascii import hexlify
 import threading
+import traceback
 import socketserver
 import logging
 import paramiko
 from paramiko.py3compat import u
 import select
+import time
+import re
+import os 
+from utils.utils import truncate_utf8_chars
 
 PORT = 3000
 REMOTE_PORT = 22
-LOG_FILE = 'sshmitm.log'
+#LOG_FILE_CLIENT = 'log/sshmitm'+time.strftime("%Y%m%d-%H%M%S")+'-CLIENT'+'.log'
+LOG_FILE_LOGIN='log/sshLogin.log'
 DENY_ALL = False
 DOMAIN = "sshd"
 # setup logging
-logger = logging.getLogger("access.log")
-logger.setLevel(logging.INFO)
-lh = logging.FileHandler(LOG_FILE)
-logger.addHandler(lh)
+def getLogger(name,log_file):
+    logger = logging.getLogger(name+".log")
+    logger.setLevel(logging.INFO)
+    lh = logging.FileHandler(log_file)
+    lh.terminator = ""
+    logger.addHandler(lh)
+    return logger,lh
+
+logger_login,_=getLogger("login",LOG_FILE_LOGIN)
+
+
+def ttyDecode(byteChar,log_file):
+    strByte=str(byteChar)
+    if strByte=="b'\\x08\\x1b[J'": #backspace
+        truncate_utf8_chars(log_file, 1, ignore_newlines=True)
+        return ""
+
+    string=byteChar.decode()
+    # ansi_escape = re.compile(r'''
+    #     \x1B  # ESC
+    #     (?:   # 7-bit C1 Fe (except CSI)
+    #         [@-Z\\-_]
+    #     |     # or [ for CSI, followed by a control sequence
+    #         \[
+    #         [0-?]*  # Parameter bytes
+    #         [ -/]*  # Intermediate bytes
+    #         [@-~]   # Final byte
+    #     )
+    # ''', re.VERBOSE)
+    #print(string)
+    badCharacter=['\[[0-9][a-z]','\[[0-9];[0-9][0-9][a-z]','\[[a-z]']
+    for regex in badCharacter:
+        ansi_escape=re.compile(regex)
+        if ansi_escape.search(string):
+            string = ansi_escape.sub('', string)
+        
+    return string
 
 
 host_key = paramiko.RSAKey(filename='test_rsa.key')
@@ -29,6 +68,7 @@ class Server (paramiko.ServerInterface):
     def __init__(self, client_address):
         self.event = threading.Event()
         self.client_address = client_address
+           
 
     def check_channel_request(self, kind, chanid):
         if kind == 'session':
@@ -36,8 +76,10 @@ class Server (paramiko.ServerInterface):
         return paramiko.OPEN_FAILED_ADMINISTRATIVELY_PROHIBITED
 
     def check_auth_password(self, username, password):
-        logger.info('IP: %s, User: %s, Password: %s' % (self.client_address[0],
-                                                        username, password))
+#        self.logger_client.info('IP: %s, User: %s, Password: %s \r\n' % (self.client_address[0],username, password))
+        logger_login.info('IP: %s, User: %s, Password: %s, Time : %s \r\n' % (self.client_address[0],
+                                                        username, password,time.strftime('%Y%m%d-%H%M%S',
+                                                        time.localtime())))                                                
         if DENY_ALL is True:
             return paramiko.AUTH_FAILED
         self.password = password
@@ -73,26 +115,29 @@ class SSHHandler(socketserver.StreamRequestHandler):
                 return
             print('Authenticated!')
 
+
             self.client = paramiko.SSHClient()
             self.client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
             self.client.connect(DOMAIN, username=server.username,
                                 password=server.password, port=REMOTE_PORT)
             chan2 = self.client.invoke_shell()
 
+            self.LOG_FILE_SERVER = 'log/sshmitm'+time.strftime('%Y%m%d-%H%M%S', time.localtime())+'-SERVER'+'.log'    
+            self.logger_server,self.lh=getLogger("server",self.LOG_FILE_SERVER)
             while True:
                 r, w, e = select.select([chan2, chan], [], [])
                 if chan in r:
                     x = chan.recv(1024)
                     if len(x) == 0:
                         break
-                    logger.info('%s' % x)
+#                    server.logger_client.info('%s' % ttyDecode(x,LOG_FILE_CLIENT))
                     chan2.send(x)
 
                 if chan2 in r:
                     x = chan2.recv(1024)
                     if len(x) == 0:
                         break
-                    logger.info('%s' % x)
+                    self.logger_server.info('%s' % ttyDecode(x,self.LOG_FILE_SERVER))
                     chan.send(x)
 
             server.event.wait(10)
@@ -105,9 +150,11 @@ class SSHHandler(socketserver.StreamRequestHandler):
 
         except Exception as e:
             print('*** Caught exception: ' + str(e.__class__) + ': ' + str(e))
+            traceback.print_exc()
         finally:
             try:
                 t.close()
+                self.logger_server.removeHandler(self.lh)
             except:
                 pass
